Se modifico albums.jsx

+++ src/components/Album/Albums.jsx
@@ -15,9 +15,10 @@ import {
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
 import PopupAlbum from "../popup/PopupAlbum";
-import { useAuth } from "../../contexts/AuthContext";
+import {useAuth} from "../../contexts/AuthContext";
 import PopupMsj from "../popup/PopupMsj";
 import PopupCreateAlbum from "../popup/PopupCreateAlbum";
+import PopupViewAlbum from "../popup/PopupViewAlbum";
 
 export default function Albums() {
   const navigate = useNavigate();
@@ -30,21 +31,26 @@ export default function Albums() {
   const [showPopupCreate, setShowPopupCreate] = useState(false);
   const [message, setMessage] = useState("");
   const [selectedAlbum, setSelectedAlbum] = useState();
+  const [filters, setFilters] = useState({});
+  const [page, setPage] = useState(1);
+  const [nextUrl, setNextUrl] = useState(null);
+  const [idAlbumDelete, setIdAlbumDelete] = useState();
+  const [showPopupAlter, setShowPopupAlter] = useState(false);
+  const [showPopupViewAlbum, setShowPopupViewAlbum] = useState(false);
+
   const handleLogout = () => {
     // Lógica para cerrar sesión
     navigate("/");
   };
   // Definir las rutas de menú manualmente si no están bajo un padre común
   const menuItems = [
-    {path: "/albums", label: "Album"},
-    {path: "/songs", label: "Canciones"},
-    {path: "/artists", label: "Artistas"},
-  //  {path: "/BuscadorDeCanciones", label: "BuscadorDeCanciones"},
+    {key: 1, path: "/albums", label: "Album"},
+    {key: 2, path: "/songs", label: "Canciones"},
+    {key: 3, path: "/ArtistList", label: "Artistas"},
+    //  {path: "/BuscadorDeCanciones", label: "BuscadorDeCanciones"},
   ];
- //para reveer
   const addAlbum = () => {
-    console.log("entro a addalbum")
-   setShowPopupCreate(true);
+    setShowPopupCreate(true);
   };
   function handleSearch(event) {
     event.preventDefault();
@@ -52,7 +58,6 @@ export default function Albums() {
     const searchForm = new FormData(event.target);
 
     const newFilters = {};
-
     searchForm.forEach((value, key) => {
       if (value) {
         newFilters[key] = value;
@@ -60,10 +65,39 @@ export default function Albums() {
     });
 
     setFilters(newFilters);
-    setSongs([]);
+    setAlbums([]);
     setPage(1);
   }
+
+  const doFetch = async () => {
+    setIsLoading(true);
+    let query = new URLSearchParams({
+      page: page,
+      page_size: 5,
+      ordering: `-created_at`,
+      ...filters,
+    }).toString();
+    fetch(`${import.meta.env.VITE_API_BASE_URL}harmonyhub/albums/?${query}`, {})
+      .then((response) => response.json())
+      .then((data) => {
+        if (data.results) {
+          setAlbums((prevAlbums) => [...prevAlbums, ...data.results]);
+          setNextUrl(data.next);
+        }
+      })
+      .catch(() => {
+        setIsError(true);
+      })
+      .finally(() => {
+        setIsLoading(false);
+      });
+  };
   useEffect(() => {
+    doFetch();
+  }, [page, filters]);
+
+  useEffect(() => {
+    setShowPopupAlter(false);
     const fetchAlbumsAndArtists = async () => {
       setIsLoading(true);
       try {
@@ -105,52 +139,68 @@ export default function Albums() {
         setIsLoading(false);
       }
     };
-
     fetchAlbumsAndArtists();
   }, [state.token]);
 
   const handleEdit = (album) => {
-    console.log("album", album);
     setSelectedAlbum(album);
     setShowPopupAlbum(true);
   };
 
   const handleDelete = (id) => {
-    console.log(`Delete album with id: ${id}`);
-    fetch(`${import.meta.env.VITE_API_BASE_URL}harmonyhub/albums/${id}/`, {
-      method: "DELETE",
-      headers: {
-        accept: "application/json",
-        Authorization: `Token ${state.token}`,
-      },
-    }).then((response) => {
-      if (!response.ok) {
-        setShowPopup(true);
-        setMessage("Las credenciales de autenticación no se proveyeron.");
-      } else {
-        setShowPopup(true);
-        setMessage("El album fue eliminado.");
-      }
-    });
+    setShowPopupAlter(false);
+    setIdAlbumDelete(id);
+    setShowPopup(true);
+    setMessage("¿Seguro que desea eliminar el album?.");
   };
 
   const handleView = (album) => {
     console.log(`View album with id: ${album}`);
-    setShowPopupAlbum(true);
-    // Implementa la lógica de visualización aquí
+    setSelectedAlbum(album);
+    setShowPopupViewAlbum(true);
   };
+
+  function handleConfigPopup() {
+    if (!showPopupAlter) {
+      fetch(
+        `${
+          import.meta.env.VITE_API_BASE_URL
+        }harmonyhub/albums/${idAlbumDelete}/`,
+        {
+          method: "DELETE",
+          headers: {
+            accept: "application/json",
+            Authorization: `Token ${state.token}`,
+          },
+        }
+      ).then((response) => {
+        if (!response.ok) {
+          setShowPopup(true);
+          setMessage("Sólo el propietario puede eliminar este álbum.");
+          setShowPopupAlter(true);
+        } else {
+          setShowPopup(true);
+          setMessage("El album fue eliminado.");
+          setShowPopupAlter(true);
+        }
+      });
+    } else {
+      setShowPopup(false);
+    }
+  }
 
   function handleClosePopup() {
     setShowPopup(false);
   }
-
-  function handlePopupCreate (){
-    console.log("entro a handlePopupCreate ")
-    setShowPopupCreate (true);
-  }
-
   function handleClosePopupAlbum() {
     setShowPopupAlbum(false);
+  }
+
+  function handleClosePopupCreate() {
+    setShowPopupCreate(false);
+  }
+  function handleClosePopupViewAlbum() {
+    setShowPopupViewAlbum(false);
   }
 
   return (
@@ -172,100 +222,103 @@ export default function Albums() {
             display: "flex",
             justifyContent: "space-between",
             alignItems: "center",
+            marginBottom: "10px",
           }}
         >
-          <h1 style={{ color: "#61F2B1", fontSize: "24px", fontWeight: "bold"}}>Albums</h1>
-          <button style={{  marginTop: "20px", padding: "10px 20px", backgroundColor: "#333", color: "#61F2B1", border: "none",  borderRadius: "5px", cursor: "#61F2B1", alignSelf: "flex-end", border: "2px solid black", }} onClick={handleLogout}>Cerrar Sesión</button>
+          <h1
+            style={{
+              color: "#61F2B1",
+              fontSize: "40px",
+              fontWeight: "bold",
+              marginLeft: "15px",
+            }}
+          >
+            Albums
+          </h1>
+          <button
+            style={{
+              marginTop: "20px",
+              padding: "10px 20px",
+              backgroundColor: "#333",
+              color: "#61F2B1",
+              border: "none",
+              borderRadius: "5px",
+              cursor: "#61F2B1",
+              alignSelf: "flex-end",
+              border: "2px solid black",
+            }}
+            onClick={handleLogout}
+          >
+            Cerrar Sesión
+          </button>
         </header>
-        <form action="">
-        <fieldset>
-    <div style={{ flex: 1, marginRight: "10px", color: "#61F2B1" }}>
-    <label>Título: 
-      <input
-        type="text"
-        name="title"
-        style={{
-          marginLeft: "10px",
-          width: "calc(100% - 20px)",
-          padding: "5px",
-          border: "1px solid black",
-        }}
-      />
-      </label>
-    <label>Artista:
-      <input
-        type="text"
-        name="artist"
-        style={{
-          marginLeft: "10px",
-          width: "calc(100% - 20px)",
-          padding: "5px",
-          border: "1px solid black",
-        }}
-      />
-    </label>
-    <label > Año:
-      <input
-        type="text"
-        name="year"
-        style={{
-          marginLeft: "10px",
-          width: "calc(100% - 20px)",
-          padding: "5px",
-          border: "1px solid black",
-        }}
-      />
-    </label>
-    <label >
-      Creado por:
-      <input
-        type="text"
-        name="createdBy"
-        style={{
-          marginLeft: "10px",
-          width: "calc(100% - 20px)",
-          padding: "5px",
-          border: "1px solid black",
-        }}
-      />
-    </label>
-    </div>
-    <div style={{ display: "flex", justifyContent: "space-between" }}>
-    <button
-      onClick={handleSearch}
-      style={{
-        marginTop: "20px",
-        padding: "10px 20px",
-        backgroundColor: "#333",
-        color: "#61F2B1",
-        border: "none",
-        borderRadius: "5px",
-        cursor: "pointer",
-        alignSelf: "flex-end",
-        border: "2px solid black",
-      }}
-    >
-      Buscar
-    </button>
-    <button onClick={addAlbum}
-        style={{
-          marginTop: "20px",
-          padding: "10px 20px",
-          backgroundColor: "#333",
-          color: "#61F2B1",
-          border: "none",
-          borderRadius: "5px",
-          cursor: "pointer",
-          alignSelf: "flex-end",
-          border: "2px solid black",
-        }}
-      >
-        Agregar album
-      </button>
-      </div>
-     
-  </fieldset>
-</form>
+        <form className="box" onSubmit={handleSearch}>
+          <div
+            style={{
+              display: "flex",
+              color: "#61F2B1",
+              justifyContent: "space-around",
+            }}
+          >
+            <div className="control">
+              <label className="label">Título:</label>
+              <input className="input" type="text" name="title" />
+            </div>
+            <div className="control">
+              <label className="label">Artista:</label>
+              <input className="input" type="text" name="<artist>" />
+            </div>
+            <div className="control">
+              <label className="label">Año:</label>
+              <input className="input" type="text" name="year" />
+            </div>
+          </div>
+          <div
+            style={{
+              display: "flex",
+              justifyContent: "space-around",
+              marginBottom: "15px",
+            }}
+          >
+            <div className="control">
+              <button
+                type="submit"
+                style={{
+                  marginTop: "20px",
+                  padding: "10px 20px",
+                  backgroundColor: "#333",
+                  color: "#61F2B1",
+                  border: "none",
+                  borderRadius: "5px",
+                  cursor: "pointer",
+                  alignSelf: "flex-end",
+                  border: "2px solid black",
+                }}
+              >
+                Buscar
+              </button>
+            </div>
+            <div className="control">
+              <button
+                type="button"
+                onClick={addAlbum}
+                style={{
+                  marginTop: "20px",
+                  padding: "10px 20px",
+                  backgroundColor: "#333",
+                  color: "#61F2B1",
+                  border: "none",
+                  borderRadius: "5px",
+                  cursor: "pointer",
+                  alignSelf: "flex-end",
+                  border: "2px solid black",
+                }}
+              >
+                Agregar album
+              </button>
+            </div>
+          </div>
+        </form>
 
         <div>
           <TableContainer component={Paper}>
@@ -331,12 +384,24 @@ export default function Albums() {
           {showPopupAlbum && (
             <PopupAlbum album={selectedAlbum} onClose={handleClosePopupAlbum} />
           )}
-          {showPopup && (
-            <PopupMsj message={message} onClose={handleClosePopup} />
+          {showPopupViewAlbum && (
+            <PopupViewAlbum
+              album={selectedAlbum}
+              onClose={handleClosePopupViewAlbum}
+            />
           )}
-          
+          {showPopup && (
+            <PopupMsj
+              message={message}
+              onConfirm={handleConfigPopup}
+              onClose={handleClosePopup}
+            />
+          )}
+          {showPopupCreate && (
+            <PopupCreateAlbum onClose={handleClosePopupCreate} />
+          )}
         </div>
       </main>
     </div>
-  ); 
+  );
 }
